--- chapter-chai/chapter_chai_initial_site_build/src/main/java/com/chapter_chai/SignInPage.css
@@ -1,51 +1,0 @@
-.page {
-    height: 100svh;
-    background-image: url("chap_chai.jpg");
-    background-position: center;
-    background-repeat: no-repeat;
-    background-size: cover;
-    width: 75vw;
-    display: flex;
-    justify-content: center;
-    text-align: center;
-}
-
-.content {
-    margin-top: 10%;
-}
-
-#title {
-    font-size: 3rem;
-}
-
-#slogan {
-    font-size: 2rem;
-}
-
-.titleCard {
-    color: #d3a680;
-}
-
-.form {
-    background-color: #f0cdb1;
-    height: 20%;
-    width: fit-content;
-    align-items: center;
-    padding: 15%;
-    padding-top: 1%;
-    border-radius: 25px;
-    margin-left: 13%;
-}
-
-.enterFields {
-    display: block;
-    margin-bottom: 10%;
-}
-
-button {
-    color: white;
-    background-color: #2c2c2c;
-    height: 30%;
-    width: 80%;
-    border-radius: 10px;
-}
