--- chapter-chai/chapter_chai_initial_site_build/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-spring.application.name=chapter_chai_initial_site_build
