+++ chapter_chai/backend/src/main/java/com/chapter_chai/webapp/controller/Controller.java
@@ -1,0 +1,15 @@
+package com.chapter_chai.webapp.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.GetMapping;
+
+
+@RestController
+@RequestMapping("/api")
+public class Controller {
+    @GetMapping("/hello")
+    public String sayHello() {
+        return "Hello!!";
+    }
+}
