+++ chapter_chai/backend/src/test/java/com/chapter_chai/webapp/WebappApplicationTests.java
@@ -1,10 +1,10 @@
-package com.chapter_chai.chapter_chai_initial_site_build;
+package com.chapter_chai.webapp;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ChapterChaiInitialSiteBuildApplicationTests {
+class WebappApplicationTests {
 
 	@Test
 	void contextLoads() {
