+++ chapter_chai/frontend/public/manifest.json
